fix(test): resolve -Wsign-compare errors under -Werror
Cast integer literals to (size_t) when comparing with size_t return
values from xTaskGroupThreads(), xTaskGroupPending(), and
std::string::size() to silence -Wsign-compare warnings treated as
errors in CI.

+++ xbase/log_test.cpp
@@ -128,7 +128,7 @@ TEST_F(LogTest, MessageTruncation) {
   xLog(false, "%s", long_msg.c_str());
 
   /* Message should be truncated to XLOG_BUF_SIZE - 1 */
-  EXPECT_EQ(captured.msg.size(), XLOG_BUF_SIZE - 1);
+  EXPECT_EQ(captured.msg.size(), (size_t)(XLOG_BUF_SIZE - 1));
   EXPECT_EQ(captured.count, 1);
 
   /* All characters should be 'A' */
